Feat: get요청 컴포넌트 위치이동 -MainPage에서 요청
